requirements.txt
streamlit>=1.30.0
groq>=0.9.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
requests>=2.28.0
